requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sentence-transformers
faiss-cpu
transformers
torch
groq
streamlit
python-dotenv